Block Culling Improvements:
Block culling now is triggered by a smaller radios around the line of sight of the camera, and then "balloons" over a second to include more blocks. It transitions similarly when there are no blocks obstructing the view.

Block Culling algorithm improved to cover more blocks with a more comfortable area in front of the player being culled

Click to Move improvements:

Right click on most intractable blocks now interacts with the block if it is within range, or moves towards the block and then immediately interacts with the block upon getting in range.

Blocks that you can interact with this way glow when both within range of both the player, and a smaller range of the mouse.

You now stop in front of entities at half entity interaction range if you target an entity, making it harder to get hurt by them while approaching.

Fixes:

Fixed blocks underneath you not necessarily being designated by the code as targeted- this hardly had an effect on gameplay and simply made consistency within programming harder